Move the private key encryption/decryption into its own module, so that the generation of new keys doesn't need to be in the main artifact.
