+++ key-encryption/pom.xml
@@ -5,9 +5,7 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>nexus-keygen</artifactId>
-
-    <name>nexus-keygen</name>
+    <artifactId>key-encryption</artifactId>
     <packaging>jar</packaging>
 
     <parent>
@@ -42,12 +40,6 @@
             <groupId>com.github.nexus</groupId>
             <artifactId>nexus-encryption-jnacl</artifactId>
             <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>key-encryption</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
 
     </dependencies>
