+++ key-encryption/src/main/java/com/github/nexus/keyenc/KeyEncryptor.java
@@ -1,4 +1,4 @@
-package com.github.nexus.keygen;
+package com.github.nexus.keyenc;
 
 import com.github.nexus.nacl.Key;
