+++ key-encryption/src/test/java/com/github/nexus/keyenc/KeyEncryptorTest.java
@@ -1,4 +1,4 @@
-package com.github.nexus.keygen;
+package com.github.nexus.keyenc;
 
 import com.github.nexus.argon2.Argon2;
 import com.github.nexus.argon2.ArgonOptions;
@@ -6,8 +6,10 @@ import com.github.nexus.argon2.ArgonResult;
 import com.github.nexus.nacl.Key;
 import com.github.nexus.nacl.NaclFacade;
 import com.github.nexus.nacl.Nonce;
+import org.assertj.core.api.Assertions;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
 
 import javax.json.Json;
 import javax.json.JsonObject;
@@ -45,22 +47,22 @@ public class KeyEncryptorTest {
         final String password = "pass";
         final ArgonResult result = new ArgonResult(new ArgonOptions("i", 1, 1, 1), new byte[]{}, new byte[]{});
 
-        doReturn(result).when(argon2).hash(eq(password), any(byte[].class));
-        doReturn(new Nonce(new byte[]{})).when(nacl).randomNonce();
+        Mockito.doReturn(result).when(argon2).hash(eq(password), any(byte[].class));
+        Mockito.doReturn(new Nonce(new byte[]{})).when(nacl).randomNonce();
         doReturn(new byte[]{}).when(nacl).sealAfterPrecomputation(any(byte[].class), any(Nonce.class), any(Key.class));
 
         final JsonObject returnJson = keyEncryptor.encryptPrivateKey(key, password);
 
         final JsonObject aopts = returnJson.getJsonObject("aopts");
 
-        assertThat(returnJson.getString("sbox")).isNotNull();
-        assertThat(returnJson.getString("asalt")).isNotNull();
-        assertThat(returnJson.getString("snonce")).isNotNull();
-        assertThat(aopts).isNotNull();
-        assertThat(aopts.getInt("memory")).isNotNull();
-        assertThat(aopts.getInt("parallelism")).isNotNull();
-        assertThat(aopts.getInt("iterations")).isNotNull();
-        assertThat(aopts.getString("variant")).isNotNull();
+        Assertions.assertThat(returnJson.getString("sbox")).isNotNull();
+        Assertions.assertThat(returnJson.getString("asalt")).isNotNull();
+        Assertions.assertThat(returnJson.getString("snonce")).isNotNull();
+        Assertions.assertThat(aopts).isNotNull();
+        Assertions.assertThat(aopts.getInt("memory")).isNotNull();
+        Assertions.assertThat(aopts.getInt("parallelism")).isNotNull();
+        Assertions.assertThat(aopts.getInt("iterations")).isNotNull();
+        Assertions.assertThat(aopts.getString("variant")).isNotNull();
 
         verify(argon2).hash(eq(password), any(byte[].class));
         verify(nacl).randomNonce();
@@ -90,13 +92,13 @@ public class KeyEncryptorTest {
             .when(nacl)
             .openAfterPrecomputation(any(byte[].class), any(Nonce.class), any(Key.class));
 
-        doReturn(new ArgonResult(null, new byte[]{}, new byte[]{4, 5, 6}))
+        Mockito.doReturn(new ArgonResult(null, new byte[]{}, new byte[]{4, 5, 6}))
             .when(argon2)
             .hash(any(ArgonOptions.class), eq(password), any(byte[].class));
 
         final Key key = keyEncryptor.decryptPrivateKey(parsed, password);
 
-        assertThat(key.getKeyBytes()).isEqualTo(new byte[]{1, 2, 3});
+        Assertions.assertThat(key.getKeyBytes()).isEqualTo(new byte[]{1, 2, 3});
 
         verify(nacl).openAfterPrecomputation(any(byte[].class), any(Nonce.class), any(Key.class));
         verify(argon2).hash(any(ArgonOptions.class), eq(password), any(byte[].class));
