+++ nexus-app/src/main/resources/nexus-spring.xml
@@ -94,7 +94,7 @@
 
     <bean name="argon2" class="com.github.nexus.argon2.Argon2" factory-method="create" />
 
-    <bean name="keyEncryptor" class="com.github.nexus.keygen.KeyEncryptorImpl" >
+    <bean name="keyEncryptor" class="com.github.nexus.keyenc.KeyEncryptorImpl" >
         <constructor-arg ref="argon2"/>
         <constructor-arg ref="nacl"/>
     </bean>
