+++ nexus-app/src/test/java/com/github/nexus/key/KeyManagerTest.java
@@ -3,7 +3,7 @@ package com.github.nexus.key;
 import com.github.nexus.TestConfiguration;
 import com.github.nexus.configuration.Configuration;
 import com.github.nexus.configuration.model.KeyData;
-import com.github.nexus.keygen.KeyEncryptor;
+import com.github.nexus.keyenc.KeyEncryptor;
 import com.github.nexus.nacl.Key;
 import com.github.nexus.nacl.KeyPair;
 import org.junit.Before;
