+++ nexus-keygen/src/main/java/com/github/nexus/keygen/KeyGeneratorImpl.java
@@ -1,6 +1,7 @@
 package com.github.nexus.keygen;
 
 import com.github.nexus.configuration.Configuration;
+import com.github.nexus.keyenc.KeyEncryptor;
 import com.github.nexus.nacl.KeyPair;
 import com.github.nexus.nacl.NaclFacade;
 import org.slf4j.Logger;
