+++ nexus-keygen/src/test/java/com/github/nexus/keygen/KeyGeneratorTest.java
@@ -1,6 +1,7 @@
 package com.github.nexus.keygen;
 
 import com.github.nexus.configuration.Configuration;
+import com.github.nexus.keyenc.KeyEncryptor;
 import com.github.nexus.nacl.Key;
 import com.github.nexus.nacl.KeyPair;
 import com.github.nexus.nacl.NaclFacade;
